pyproject.toml
[project]
name = "secret-ai-writer"
version = "0.1.0"
description = ""
authors = [
    {name = "emlanis", email = "emlanis@scrt.network"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "python-decouple (>=3.8,<4.0)",
    "langchain-core (>=0.1.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
secret-sdk = "^1.8.1"
secret-ai-sdk = "*"